refactor: move record-metadata component from datahub to feature/record lib
This way the datahub app holds less presentation logic
