+++ apps/datahub/src/app/app.component.html
@@ -33,6 +33,7 @@
     <gn-ui-main-search
       *ngIf="(isRecordOpened$ | async) === false"
     ></gn-ui-main-search>
-    <gn-ui-record-view *ngIf="isRecordOpened$ | async"> </gn-ui-record-view>
+    <gn-ui-record-metadata *ngIf="isRecordOpened$ | async">
+    </gn-ui-record-metadata>
   </div>
 </main>
