--- apps/datahub/src/app/dataset-view/record-view.component.html
@@ -1,53 +1,0 @@
-<mat-tab-group (selectedIndexChange)="onTabIndexChange($event)">
-  <mat-tab>
-    <ng-template mat-tab-label>
-      <mat-icon class="mr-3">info</mat-icon>
-      <span translate="">mdview.tab.info</span>
-    </ng-template>
-    <gn-ui-metadata-info
-      *ngIf="(facade.isPresent$ | async) === false"
-      [metadata]="{
-        title: 'Loading...'
-      }"
-      [incomplete]="true"
-    >
-    </gn-ui-metadata-info>
-    <gn-ui-metadata-info
-      *ngIf="(facade.isPresent$ | async) === true"
-      [metadata]="facade.metadata$ | async"
-      [incomplete]="facade.isIncomplete$ | async"
-    >
-    </gn-ui-metadata-info>
-  </mat-tab>
-
-  <mat-tab>
-    <ng-template mat-tab-label>
-      <mat-icon class="mr-3">list</mat-icon>
-      <span translate="">mdview.tab.data</span>
-    </ng-template>
-  </mat-tab>
-
-  <mat-tab [disabled]="(displayMap$ | async) === false">
-    <ng-template mat-tab-label>
-      <mat-icon class="mr-3">map</mat-icon>
-      <span translate="">mdview.tab.map</span>
-    </ng-template>
-    <div style="height: 500px" *ngIf="displayMap$ | async">
-      <gn-ui-data-view-map></gn-ui-data-view-map>
-    </div>
-  </mat-tab>
-
-  <mat-tab>
-    <ng-template mat-tab-label>
-      <mat-icon class="mr-3">download</mat-icon>
-      <span translate="">mdview.tab.exports</span>
-    </ng-template>
-    <gn-ui-data-downloads></gn-ui-data-downloads>
-  </mat-tab>
-  <mat-tab>
-    <ng-template mat-tab-label>
-      <mat-icon class="mr-3">api</mat-icon>
-      <span translate="">mdview.tab.api</span>
-    </ng-template>
-  </mat-tab>
-</mat-tab-group>
