+++ libs/feature/record/src/lib/record-metadata/record-metadata.component.ts
@@ -1,4 +1,7 @@
-import { Component, ChangeDetectionStrategy } from '@angular/core'
+import { ChangeDetectionStrategy, Component } from '@angular/core'
+import { MapManagerService } from '@geonetwork-ui/feature/map'
+import { MdViewFacade } from '../state/mdview.facade'
+import { map } from 'rxjs/operators'
 
 @Component({
   selector: 'gn-ui-record-metadata',
@@ -6,4 +9,19 @@ import { Component, ChangeDetectionStrategy } from '@angular/core'
   styleUrls: ['./record-metadata.component.css'],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class RecordMetadataComponent {}
+export class RecordMetadataComponent {
+  displayMap$ = this.facade.mapApiLinks$.pipe(
+    map((links) => !!links && links.length > 0)
+  )
+
+  constructor(
+    public facade: MdViewFacade,
+    private mapManager: MapManagerService
+  ) {}
+
+  onTabIndexChange(index: number): void {
+    if (index === 2) {
+      setTimeout(() => this.mapManager.map.updateSize(), 0)
+    }
+  }
+}
